+++ ft_strlcat.c
@@ -10,7 +10,9 @@
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlcat(char *dest, char *src, int destSize)
+#include "libft.h"
+
+size_t	ft_strlcat(char *dest, char *src, size_t destSize)
 {
 	int	i;
